rainbow-vibe/requirements.txt
boto3
bedrock-agentcore
strands-agents 
strands-agents-tools
bedrock-agentcore-starter-toolkit
uv
playwright
nest_asyncio
streamlit